docs: Updating the chapter references in step 4
The Chapters references mentioned refer to the second edition
of the Speech and Language Processing textbook, while the provided link
refers to the current edition, the third edition, which have been restructured.
In the third edition, as of December 30, 2020 draft release
the following changes have been amended:
Chapter 4 of the 2nd ed. is now Chapter 3 of the 3rd ed. ,
Chapter 9 of the 2nd ed. is now Chapter 26 of the 3rd ed. ,
Chapter 10 of the 2nd ed. content have been discarded.